+++ src/main/java/com/cm/welfarecmcity/api/loan/LoanService.java
@@ -55,7 +55,7 @@ public class LoanService {
     //loanDto.setLoanNo("2566-0624001");
     loanDto.setLoanValue(Double.parseDouble(req.getLoanValue()));
     loanDto.setLoanBalance(Double.parseDouble(req.getLoanValue()));
-    loanDto.setLoanTime(req.getLoanTime());
+    loanDto.setLoanTime(Long.valueOf(req.getLoanTime()).intValue());
     loanDto.setInterest(Integer.parseInt(req.getInterestLoan()));
     loanDto.setInterestPercent(Integer.parseInt(req.getInterestPercent()));
     loanDto.setNewLoan(true);
